err handlers api

+++ opinions_app/api_views.py
@@ -3,29 +3,35 @@
 from flask import jsonify, request
 
 from . import app, db
+from .error_handlers import InvalidAPIUsage
 from .models import Opinion
 from .views import random_opinion
 
 
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом.
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     return jsonify({'opinion': opinion.to_dict()}), 200
 
 
 @app.route('/api/opinions/<int:id>/', methods=['PATCH'])
 def update_opinion(id):
     data = request.get_json()
+    if 'title' and 'text' not in data:
+        raise InvalidAPIUsage(
+            'В запросе отсутствуют корретные поля для измненения')
     if (
         'text' in data and 
         Opinion.query.filter_by(text=data['text']).first() is not None
     ):
-        # При неуникальном значении поля text
-        # возвращаем сообщение об ошибке в формате JSON
-        # и статус-код 400:
-        return jsonify({'error':
-                        'Такое мнение уже есть в базе данных'}), 400
-    opinion = Opinion.query.get_or_404(id)
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
+    opinion = Opinion.query.get(id)
+    # Тут код ответа нужно указать явным образом.
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     opinion.title = data.get('title', opinion.title)
     opinion.text = data.get('text', opinion.text)
     opinion.source = data.get('source', opinion.source)
@@ -36,7 +42,10 @@ def update_opinion(id):
 
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        # Тут код ответа нужно указать явным образом.
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     return '', 204
@@ -53,23 +62,20 @@ def get_opinions():
 def add_opinion():
     data = request.get_json()
     if 'title' not in data or 'text' not in data:
-        return jsonify({'error':
-                        'В запросе отсутствуют обязательные поля'}), 400
-    # Если в базе данных уже есть объект 
-    # с таким же значением поля text...
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля')
     if Opinion.query.filter_by(text=data['text']).first() is not None:
-        # ...возвращаем сообщение об ошибке в формате JSON 
-        # и статус-код 400:
-        return jsonify({'error': 
-                        'Такое мнение уже есть в базе данных'}), 400
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
     opinion = Opinion()
     opinion.from_dict(data)
     db.session.add(opinion)
     db.session.commit()
-    return jsonify({'opinion': opinion.to_dict()}), 201 
+    return jsonify({'opinion': opinion.to_dict()}), 201
 
 
 @app.route('/api/get-random-opinion/', methods=['GET'])
 def get_random_opinion():
     opinion = random_opinion()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    # Тут код ответа нужно указать явным образом.
+    raise InvalidAPIUsage('В базе данных нет мнений', 404)
